backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
asttokens==3.0.0
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.2
bs4==0.0.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.17
decorator==5.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fonttools==4.60.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
idna==3.11
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.1.0
redis==7.0.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
